% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{ch_dependency}
\alias{ch_dependency}
\title{codehover HTML dependency}
\usage{
ch_dependency()
}
\value{
An [htmltools::htmlDependency()] object.
}
\description{
Returns the htmltools dependency (CSS + JavaScript) that
  makes codehover tables interactive. It is attached automatically by
  [ch_hover()] and [ch_out()], so you rarely need to call this yourself.
  Useful if you build the HTML manually and only want the assets.
}
