% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipbookr_reduce.R
\name{chunk_code_hover}
\alias{chunk_code_hover}
\title{refer to chunk that only has graphical output at every pause point
Build code hover, figures must be built separately at this point}
\usage{
chunk_code_hover(chunk_name, fig_path_and_prefix)
}
\arguments{
\item{fig_path_and_prefix}{}
}
\value{

}
\description{
refer to chunk that only has graphical output at every pause point
Build code hover, figures must be built separately at this point
}
