% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipbookr_reduce.R
\name{chunk_prepare_assets}
\alias{chunk_prepare_assets}
\title{Title}
\usage{
chunk_prepare_assets(
  chunk_name = "example_name",
  fig.path = paste0("assets_ch_", chunk_name, "/"),
  fig_prefix = "figure"
)
}
\arguments{
\item{fig_prefix}{}
}
\value{

}
\description{
Title
}
