% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipbookr_reduce.R
\name{partial_code_ch_format}
\alias{partial_code_ch_format}
\title{prepare partial code for code hover from flipbookr partial code list}
\usage{
partial_code_ch_format(partial_code_list, which)
}
\arguments{
\item{partial_code_list}{full partial code sequence}

\item{which}{points to element in list}
}
\value{
partial code formatted in the code hover way.
}
\description{
prepare partial code for code hover from flipbookr partial code list
}
