% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipbookr_reduce.R
\name{partial_code_ch_int_reduce}
\alias{partial_code_ch_int_reduce}
\title{Produces the pipeline needed for hover reveal from partial code list ()}
\usage{
partial_code_ch_int_reduce(partial_code_list, fig_path_and_prefix)
}
\arguments{
\item{partial_code_list}{flipbookr partial code sequences as a list}

\item{fig_path_and_prefix}{paths to figures (png), minus numbering and extensions ".png"}
}
\value{

}
\description{
Produces the pipeline needed for hover reveal from partial code list ()
}
