% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ch_int}
\alias{ch_int}
\title{CodeHover initiate Table}
\usage{
ch_int(type = "incremental", css_class = "", ...)
}
\arguments{
\item{type}{type of hover effect, options are "one_row" or "incremental", the later highlights every prior row and the row curent hovered. Default is "incremental".}

\item{css_class}{Pass a CSS class name here as text to style the whole table.}

\item{...}{Adicional paramethers}
}
\value{
The beggining of an HTML table to be passed to the function \code{ch_row}
}
\description{
CodeHover initiate Table
}
