% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHover_functions.R
\name{ch_int}
\alias{ch_int}
\title{codehover initiate Table}
\usage{
ch_int(
  type = "incremental",
  css_class = "",
  table_tag_add = "",
  div_tag_add = "",
  ...
)
}
\arguments{
\item{type}{(string) Type of hover effect, options are "one_row" or "incremental", the later highlights every prior row and the row curent hovered. Default is "incremental".}

\item{css_class}{(string) Pass a CSS class name here as text to style the whole table.}

\item{table_tag_add}{(string) Aditional html parameter for table tag.}

\item{div_tag_add}{(string) Aditional html parameter for div tag (table container).}

\item{...}{Adicional paramethers}
}
\value{
The beggining of an HTML table to be passed to the function \code{ch_row}
}
\description{
Codehover initiate HTML table. Here you can adjust some HTML-style properties, and choose the type of hover effect.
}
