% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHover_functions.R
\name{ch_out}
\alias{ch_out}
\title{codehover output}
\usage{
ch_out(
  .data = "",
  img = "",
  css_class = "",
  url = FALSE,
  img_tag_add = "",
  div_tag_add = "",
  ...
)
}
\arguments{
\item{.data}{codehover object (string of text) created by ch_row.}

\item{img}{Image to be show in the image holder prior to any hover interaction.If url = TRUE the img should be a url link to a image.}

\item{css_class}{Pass a CSS class name here as text to style the image holder.}

\item{url}{Default FALSE, if TRUE indicates that the image is hosted somewhere else and it should not be embeded in the HTML page via base64.}

\item{img_tag_add}{string aditional html parameter for img tag}

\item{div_tag_add}{string aditional html parameter for div tag (container)}

\item{...}{Adicional paramethers}
}
\value{
A HTML table as string ready to be passed to \code{htmltools::HMTL} and be render in the page.
}
\description{
codehover define image holder linked to the table
}
