% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ch_out}
\alias{ch_out}
\title{CodeHover output}
\usage{
ch_out(.data = "", img_holder = "", img_holder_css_class = "", ...)
}
\arguments{
\item{.data}{CodeHover object (string of text) created by ch_row.}

\item{img_holder}{Image to be show in the image holder prior to any hover.}

\item{img_holder_css_class}{Pass a CSS class name here as text to style the image holder.}

\item{...}{Adicional paramethers}
}
\value{
A HTML table as string ready to be passed to \code{htmltools::HMTL} and be render in the page.

\dontrun{
ch_out(img_holder = "./IMG/image1.png")
}
}
\description{
CodeHover define image holder linked to the table
}
