% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ch_row}
\alias{ch_row}
\title{CodeHover add row}
\usage{
ch_row(.data = "", text = "", img = "", url = FALSE, ...)
}
\arguments{
\item{.data}{CodeHover object (string of text) created by ch_int.}

\item{text}{Text to be display in the table row.}

\item{img}{Path to the image. If url = TRUE the img should be a url link to a image.}

\item{url}{Default FALSE, if TRUE indicates that the image is hosted somewhere else and it should not be embeded in the HTML page via base64.}

\item{...}{Adicional paramethers}
}
\value{
The populated HTML table to be passed to the function \code{ch_out}

\dontrun{
ch_row(text = "Some code-text in a row",
      img = "./IMG/image1.png")
}
}
\description{
CodeHover add row to the table and images to be linked to each row
}
