% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipbookr_reduce.R
\name{chunk_code_hover}
\alias{chunk_code_hover}
\title{refer to chunk that only has graphical output at every pause point
Build code hover, figures must be built separately at this point}
\usage{
chunk_code_hover(
  chunk_name,
  fig.path = paste0("assets_ch_", chunk_name, "/"),
  fig_prefix = "figure",
  fig_path_and_prefix = past0(fig.path, fig_prefix)
)
}
\arguments{
\item{chunk_name}{refer to a code chunk by name}

\item{fig_path_and_prefix}{paths to figures (png), minus numbering and extensions ".png"}
}
\value{

}
\description{
refer to chunk that only has graphical output at every pause point
Build code hover, figures must be built separately at this point
}
